internal: test reliability improvements
* Switch from shutting down gRPC server with `Stop()` to `GracefulStop()`.
  which waits for current requests to drain before returning. This
  means that tests which log to the test runner won't ever attempt to
  log after the test has completed.

* Hoist test loggers into the fixtures package and use them from there.

* Add a workgroup context cancellation test.

This updates #2775.
