+++ internal/featuretests/featuretests.go
@@ -31,6 +31,7 @@ import (
 	"github.com/projectcontour/contour/internal/assert"
 	"github.com/projectcontour/contour/internal/contour"
 	"github.com/projectcontour/contour/internal/dag"
+	"github.com/projectcontour/contour/internal/fixture"
 	cgrpc "github.com/projectcontour/contour/internal/grpc"
 	"github.com/projectcontour/contour/internal/k8s"
 	"github.com/projectcontour/contour/internal/metrics"
@@ -38,7 +39,6 @@ import (
 	"github.com/projectcontour/contour/internal/sorter"
 	"github.com/projectcontour/contour/internal/workgroup"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/tools/cache"
@@ -54,17 +54,10 @@ const (
 	statsPort    = 8002
 )
 
-type discardWriter struct{}
-
-func (d *discardWriter) Write(buf []byte) (int, error) {
-	return len(buf), nil
-}
-
 func setup(t *testing.T, opts ...interface{}) (cache.ResourceEventHandler, *Contour, func()) {
 	t.Parallel()
 
-	log := logrus.New()
-	log.Out = new(discardWriter)
+	log := fixture.NewTestLogger(t)
 
 	et := &contour.EndpointsTranslator{
 		FieldLogger: log,
@@ -118,21 +111,17 @@ func setup(t *testing.T, opts ...interface{}) (cache.ResourceEventHandler, *Cont
 
 	l, err := net.Listen("tcp", "127.0.0.1:0")
 	check(t, err)
-	discard := logrus.New()
-	discard.Out = new(discardWriter)
 	// Resource types in xDS v2.
-	srv := cgrpc.NewAPI(discard, append(contour.ResourcesOf(resources), et), r)
+	srv := cgrpc.NewAPI(log, append(contour.ResourcesOf(resources), et), r)
 
 	var g workgroup.Group
 
 	g.Add(func(stop <-chan struct{}) error {
-		done := make(chan error)
 		go func() {
-			done <- srv.Serve(l) // srv now owns l and will close l before returning
+			<-stop
+			srv.GracefulStop()
 		}()
-		<-stop
-		srv.Stop()
-		return <-done
+		return srv.Serve(l) // srv now owns l and will close l before returning
 	})
 	g.Add(eh.Start())
